refactor(scheduler): extract pending-restore registry types to scheduler_pending_restore.py
Phase B of the scheduler decomposition extraction port.

Move the module-level fork helpers — PENDING_RESTORE_REGISTRY_MAX
constant and PendingRestoreEntry dataclass — out of monolithic
scheduler.py into a new fork-only file
python/sglang/srt/managers/scheduler_pending_restore.py. Aligns with
upstream's pattern of moving module-level helpers out of scheduler.py
(upstream PR #25638 moved EmbeddingBatchResult similarly).

scheduler.py re-exports the names at module scope via a top-level
import, preserving the existing public API. No test changes needed;
test_pending_restore_registry.py continues to import from the
scheduler module without modification.

Marker policy: header-only on the new file (Apache license header
followed by ENGRAM_MODIFIED line, no internal BEGIN/END blocks) per
the snapshot.py precedent for fully fork-authored files. The
scheduler.py import is wrapped in a clean BEGIN/END ENGRAM block.

Semantics-preserving — no behavior change. Block balance unchanged
(15 BEGIN / 14 END, with the pre-existing L29 orphan END and L30
unclosed BEGIN slated for Phase F hygiene).

py_compile passes on both touched files.
